test: cqlpy: host the single-node strongly consistent tests
The pytest coverage for strongly consistent tables all sits in
test/cluster, because the feature is experimental and the cqlpy suite
shares one server, so enabling it there enables it for every test. Two
of those tests never needed more than one node and plain CQL, yet each
of them boots and discards a cluster of its own.

Give them a test/cqlpy/strong_consistency directory with a
test_config.yaml of its own - the suite lookup walks up from the test
file, so the directory gets a server with the feature on while the rest
of cqlpy keeps the server it has today. Enabling it for the whole suite
would have been a smaller change, but it stops tablet balancing
cluster-wide, and test/alternator/run builds its command line from
test/cqlpy/run.py and has tests which resize tablets. Under
test/cqlpy/run, which serves the whole directory from one server, the
keyspace fixture skips. The directory only exists for that reason: once
the feature stops disabling tablet balancing, the test file should move
up into test/cqlpy.
