+++ test/cqlpy/strong_consistency/__init__.py
@@ -1,0 +1,5 @@
+#
+# Copyright (C) 2026-present ScyllaDB
+#
+# SPDX-License-Identifier: LicenseRef-ScyllaDB-Source-Available-1.1
+#
